requirements.txt
# Core Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# OpenAI / LLM
openai>=1.3.0

# Database
sqlalchemy>=2.0.0

# Development
httpx>=0.25.0  # For testing

# Optional: Production additions
# gunicorn>=21.2.0  # Production WSGI server
# psycopg2-binary>=2.9.9  # For PostgreSQL
# redis>=5.0.0  # For caching
# celery>=5.3.0  # For background tasks
